requirements.txt
flask
flask-cors
gunicorn
python-dotenv
langchain
langchain_openai
langchain_pinecone
openai
pinecone-client